Move flask test to comman test, update endpoints import
