+++ wb/main/api_endpoints/v1/common_test.py
@@ -6,8 +6,13 @@ Tests for common endpoints
 # pylint: disable=wrong-import-position
 from workbench.workbench import APP
 
+from flask_migrate import upgrade
+
+from migrations.migration import APP as MIGRATION_APP
+from wb import CLIToolsOptionsCache
+from wb.main.database import data_initialization
+
 from wb.extensions_factories.database import get_db_session_for_app
-from wb.flask_test import TestFlaskAppCase
 from wb.main.api_endpoints.utils import find_projects, project_json, connect_with_parents
 from wb.main.enumerates import OptimizationTypesEnum, SupportedFrameworksEnum, TargetTypeEnum
 from wb.main.models.datasets_model import DatasetsModel
@@ -18,6 +23,21 @@ from wb.main.models.topologies_metadata_model import TopologiesMetaDataModel
 from wb.main.models.topologies_model import TopologiesModel
 
 
+def apply_migrations():
+    with MIGRATION_APP.app_context():
+        upgrade()
+
+
+class TestFlaskAppCase:
+    @staticmethod
+    def setup():
+        apply_migrations()
+        data_initialization.initialize(MIGRATION_APP)
+        CLIToolsOptionsCache().initialize()
+
+
+# import pytest
+# @pytest.mark.skip(reason='Blueprints registration')
 class TestCommonApiCase(TestFlaskAppCase):
     @staticmethod
     def create_topology_record() -> TopologiesModel:
